Use https://datatables.net to render the HTML table with the recipe index
The benefit of using DataTables.net is that it renders the DOM
dynamically showing just a subset/page of the data at a time.
It also provides filtering, sorting and many more settings
(https://datatables.net/reference/option/)

Also on top I have added a small table that shows the number of recipes
with additional platforms and their ratio to the total.

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>